Update cn.json
Completed my 1st draft from top to bottom. As mentioened in Telegram, there are some items that do not make sense and required to be translated into english before can be understood what is that meant for. So, i can not just translate but convey the message across to the player in a meaningful manner.